Refactored addToCart EndPoint to return dynamic HttpStatus. Instead of always returning 200 status for both creating and updating product in the cart, now it returns 200 for updating and 201 for creating.
Added addProduct, deleteProduct and getProduct in Product Controller

+++ src/main/java/springbootlearning/ecommerce/entities/Product.java
@@ -34,9 +34,9 @@ public class Product {
     @JoinColumn(name = "category_id")
     private Category category;
 
-    public Product(String name, Double price, String description, String status){
+    public Product(String name, BigDecimal price, String description, String status){
         this.name = name;
-        this.price = BigDecimal.valueOf(price);
+        this.price = price;
         this.description = description;
         this.status = status;
     }
